Add 'type+bytes' option type
The epb_hash and if_filter options are both defined as having their
first octet be a type indicator, and the remainder being bytes data.
This adds a 'type+bytes' option type to distinguish these fields as
such.

Previously if_filter was being incorrectly treated as a string. The
official spec is incomplete in its definition of if_filter, but it's
treated in Wireshark as being either 0x00 followed by a string filter
(libpcap style), or 0x01 followed by the raw BPF bytecode.

I've also determined from the spec that epb_flags IS endianness
dependent, so I've marked it as u32 to formalize this.