Add ability to create blocks in python code and write them
The schema for a block now takes 3-tuples instead of 2-tuples, with the
third member being the default value to use. This default is used when a
dev creates a block object manually; it's completely ignored when
reading blocks from files.

I've removed `Block.from_context()` and instead created
`SectionHeader.new_member()` to handle creating blocks that must be a
member of a section. `pcapng/scanner.py` has been tweaked to use this
new method, so the behaviour for reading pcapng files should be
transparent.

Added `examples/generate_pcapng.py` to demonstrate how to create pcapng
files and blocks by hand. It creates a SHB, IDB, and SPB (from
hard-coded data) and writes them to the given filename.