Improve SimplePacket API
Added `BlockWithInterfaceMixin` to `SimplePacket`, and added the
`interface_id` property to return 0, since by definition, a SPB is
assumed to have come from that interface.

Added `captured_len` property, which accounts for the interface's
`snaplen`. This makes the API compatible with that for `EnhancedPacket`
and `Packet`.

Corrected minor logic errors in the `packet_len` and `packet_data`
properties.

Also, for all three packet block types, changed the default value for
program-created blocks' Data field to be `None` for a slight performance
boost.