Track which options are repeatable
Created a namedtuple type ``Option()`` to hold option definitions. This
allows optional elements with defaults; the new key ``multiple``
defaults to False, but can be set to True for options which explicitly
allow it (eg ``opt_comment``).

Tweaked ``Options()`` a bit, so instead of storing dicts to hold the
namee and ftype of each option code, it just stores the ``Option()`` for
the code.

Refactored some code that was affected or made redundant by the above
change.

Added a warning when loading pcapng files with repeated values for
non-repeatable options. Added warn/fix/fail for user-setting of same,
and for writing of same to file (just in case).