Add API for querying/changing `epb_flags` option
Add a FlagWord class which holds an integer and is given a schema of the
flags, enumerations, and numbers that make up that integer. Give a nice
API for getting and setting the individual parts. Encode to the file as
the integer.

Add an EPBFlags class to structs and an `epb_flags` option type to
Options.