Rework PacketBytes again
Removed the hacky need to set a PacketBytes object's capture_len
manually. Added the slightly-less hacky ability for later fields in a
schema to depend on fields that came earlier in the schema. PacketBytes
is the only one that uses this of course; you pass it the name of the
field which has the actual packet length.

That still doesn't work for SimplePacket since its length doesn't
necessarily come from a field in the block. I just ended up overriding
its _decode method instead.